@@ -10,34 +10,36 @@
   "dependencies": {
     "@babel/runtime": "^7",
     "@babel/runtime-corejs3": "^7",
-    "@mdi/font": "^4",
-    "axios": "^0.19.0",
+    "@mdi/font": "^6",
+    "axios": "^0.26.1",
     "babel-core": "^6.26.3",
     "core-js": "^3",
-    "css-loader": "^3",
+    "css-loader": "^6",
     "md-editor2": "^0.1.12",
-    "vue": "^2.6",
+    "vue": "^2.6.4",
     "vue-markdown-renderer": "^0.2.7",
     "vue-qrcode-component": "^2.1.1",
-    "vue-router": "^3.0.3",
-    "vuetify": "^2.0.0",
+    "vue-router": "^3.0.0",
+    "vuetify": "^2.6.4",
     "vuetify-google-autocomplete": "^2",
-    "vuex": "^3.0.1",
-    "webpack": "^4.39.1",
-    "webpack-dev-server": "^3.7.2"
+    "vuex": "^3.0.0",
+    "webpack": "^5.70.0",
+    "webpack-dev-server": "^4.7.4"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~5.0.0",
-    "@vue/cli-plugin-eslint": "^3.10.0",
-    "@vue/cli-service": "^3.10.0",
+    "@vue/cli-plugin-babel": "~5.0.3",
+    "@vue/cli-plugin-eslint": "^5.0.3",
+    "@vue/cli-service": "^5.0.3",
     "babel-eslint": "^10.0.1",
-    "eslint": "^5.16.0",
-    "eslint-plugin-vue": "^5.0.0",
-    "sass": "^1.17.4",
-    "sass-loader": "^7.1.0",
-    "vue-cli-plugin-vuetify": "^0.6.1",
+    "eslint": "^8.11.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-vue": "^8.5.0",
+    "sass": "^1.49.9",
+    "sass-loader": "^12.6.0",
+    "vue-cli-plugin-vuetify": "^2.4.8",
     "vue-template-compiler": "^2",
-    "vuetify-loader": "^1.2.2"
+    "vuetify-loader": "^1.7.3"
   },
   "eslintConfig": {
     "root": true,
