@@ -1,5 +1,4 @@
 export function split_carriagereturn(value) {
-  if (typeof value != "string")
-    return null;
-  return value.split("\n");
+  if (typeof value !== 'string') { return null; }
+  return value.split('\n');
 }
