@@ -6,12 +6,12 @@ Boo! Hop! Skip!
 
 <script>
 // @ is an alias to /src
-//import HelloWorld from '@/components/HelloWorld.vue'
+// import HelloWorld from '@/components/HelloWorld.vue'
 
 export default {
   name: 'home',
   components: {
-//    HelloWorld
-  }
-}
+    //    HelloWorld
+  },
+};
 </script>
