@@ -7,31 +7,28 @@ module.exports = {
       '^/concrescent/': {
         target: 'https://tsaukpaetra.com',
         changeOrigin: true, // so CORS doesn't bite us.
-      }
-    }
+      },
+    },
   },
 
-  chainWebpack: config => {
+  chainWebpack: (config) => {
     config
       .plugin('html')
-      .tap(args => {
+      .tap((args) => {
         args[0].template = './src/index.html';
         args[0].title = 'ConCrescent';
         args[0].favicon = './customization/favicon.ico';
-        return args
+        return args;
       });
     config
       .plugin('define')
-      .tap(args => {
-        return args
-      });
-
+      .tap((args) => args);
   },
   transpileDependencies: [
-        // can be string or regex
-    ],
+    // can be string or regex
+  ],
 
   pluginOptions: {
-    vuetify: {}
-  }
-}
+    vuetify: {},
+  },
+};
